+++ workflow/utils/processing.py
@@ -30,6 +30,7 @@ except ImportError as e:
 from .assertions import assert_neighbors
 from .io import to_memory, csr_matrix_int64_indptr
 from .misc import dask_compute
+from .accessors import _filter_genes
 
 
 def compute_neighbors(adata, output_type=None, force=False, check_n_neighbors=False, **kwargs):
@@ -88,44 +89,6 @@ def compute_neighbors(adata, output_type=None, force=False, check_n_neighbors=Fa
     assert_neighbors(adata, check_n_neighbors=False)
 
 
-def _filter_genes(adata, **kwargs):
-    import sparse
-    from tqdm.dask import TqdmCallback
-    import scanpy as sc
-    from dask import array as da
-    
-    if isinstance(adata.X, da.Array):
-        # adata.X = adata.X.map_blocks(
-        #     lambda x: x.toarray(),
-        #     dtype=adata.X.dtype
-        # )
-        min_cells = kwargs.get('min_cells', 0)
-        mask = adata.X.map_blocks(sparse.COO).sum(axis=0) >= min_cells
-        with TqdmCallback(
-            desc=f"Determine genes with >= {min_cells} cells",
-            miniters=10,
-            mininterval=5,
-        ):
-            mask = mask.compute().todense()
-        return mask
-    gene_subset, _ = sc.pp.filter_genes(adata.X, **kwargs)
-    # if isinstance(gene_subset, da.Array):
-    #     gene_subset = gene_subset.compute()
-    return gene_subset
-
-
-def _filter_batch(adata, batch_key=None):
-    """
-    Filter cells from batches with too few cells
-    """
-    mask = np.ones(adata.n_obs, dtype=bool)
-    if batch_key is not None:
-        cells_per_batch = adata.obs[batch_key].value_counts()
-        if cells_per_batch.min() < 2:
-            mask = adata.obs[batch_key].isin(cells_per_batch[cells_per_batch > 1].index)
-    return mask
-
-
 def filter_genes(adata, batch_key=None, compute=True, **kwargs):
     """
     Filter anndata based on .X matrix
@@ -140,10 +103,10 @@ def filter_genes(adata, batch_key=None, compute=True, **kwargs):
             adata = adata[adata.obs[batch_key].isin(cells_per_batch[cells_per_batch > 1].index)]
         
     # apply filter
-    gene_subset = _filter_genes(adata, **kwargs)
-    if any(gene_subset == False):
-        logging.info(f'Subset to {sum(gene_subset)}/{adata.n_vars} filtered genes...')
-        adata = adata[:, gene_subset]
+    gene_mask = _filter_genes(adata, **kwargs)
+    if any(gene_mask == False):
+        logging.info(f'Subset to {sum(gene_mask)}/{adata.n_vars} filtered genes...')
+        adata = adata[:, gene_mask]
     
     if adata.is_view:
         adata = adata.copy()
